% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_graph .R
\name{transition_graph}
\alias{transition_graph}
\title{Transition Analysis Graph}
\usage{
transition_graph(png, data, title)
}
\arguments{
\item{png}{Path where we want to save the file. It is a string that should end in .png. If png = FALSE, the function will plot the graph in R without saving it.}

\item{data}{Data frame containing the information to construct the graph. It needs to have these three columns at least:
"value" containing the numeric value of the main source of income express in euro per year
"category" containing the name of the main source of income
"year" containing the reference year}

\item{title}{Title of the graph}
}
\value{
This function return a plot, representing the percentage receiving a certain source of income as their main source of income over time.
}
\description{
This function plot of a transition analysis, studying the transition from a certain source of income to other source of income. We consider in particular only five sources: unemployment, pension, sick benefit, minimum wage, and earned wage.
}
